// src/src/types/request.d.ts
export interface ToolFeature {
    toolId: number;
    toolFeatureId: number;
    toolFeatureName: string;
    toolFeatureDescription: string;
    toolModuleName: string;
    accessTypeName: string;
    isActivePermission: boolean;
}

// AD group with features
export interface AdGroup {
    adGroup: string;
    adGroupId: string;
    toolFeature: ToolFeature[];
}

// Tool (under a role)
export interface IamTool {
    toolId: number;
    adGroups: AdGroup[];
}

// Role within iamRoles
export interface IamRole {
    roleId: number;
    role: string;
    roleLevelName: string;
    subFunctionName: string;
    functionID: number;
    subFunctionID: number;
    department: string;
    region: string;
    regionID: number;
    clusterID: number;
    marketID: number;
    siteID: number;
    roleType: string;
    tools: IamTool[];
}

// Full request payload
export interface DelegationRequestPayload {
    delegatedToUser?: string;
    delegatedByUser?: string;
    loggedInUser?: string;
    startDate: Date | string;
    endDate: Date | string;
    requestComment: string;
    roleIds?: string;
    toolIds: string;
    iamRoles: IamRole[];
    delegationId: number;
}